Refactor CA client code.
