+++ security/cmd/node_agent/na/nafactory.go
@@ -17,10 +17,12 @@ package na
 import (
 	"fmt"
 	"os"
+
 	// TODO(nmittler): Remove this
 	_ "github.com/golang/glog"
 
 	"istio.io/istio/pkg/log"
+	"istio.io/istio/security/pkg/caclient/grpc"
 	"istio.io/istio/security/pkg/platform"
 	"istio.io/istio/security/pkg/workload"
 )
@@ -47,7 +49,7 @@ func NewNodeAgent(cfg *Config) (NodeAgent, error) {
 		return nil, err
 	}
 
-	cAClient := &cAGrpcClientImpl{}
+	cAClient := &grpc.CAGrpcClientImpl{}
 	na.cAClient = cAClient
 
 	// TODO: Specify files for service identity cert/key instead of node agent files.
