+++ security/pkg/caclient/grpc/mock/fakeclient.go
@@ -1,0 +1,40 @@
+// Copyright 2017 Istio Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package mock
+
+import (
+	"fmt"
+
+	"istio.io/istio/security/pkg/platform"
+	pb "istio.io/istio/security/proto"
+)
+
+const (
+	maxCAClientSuccessReturns = 8
+)
+
+type FakeCAClient struct {
+	Counter  int
+	Response *pb.Response
+	Err      error
+}
+
+func (f *FakeCAClient) SendCSR(req *pb.Request, pc platform.Client, addr string) (*pb.Response, error) {
+	f.Counter++
+	if f.Counter > maxCAClientSuccessReturns {
+		return nil, fmt.Errorf("terminating the test with errors")
+	}
+	return f.Response, f.Err
+}
